Add makefile
The makefile is intended to use this repository as submodule in a bigger
project (well, this is the whole purpose of this thing, right?).

The makefile produces a static library (.a) to include it into your
project.

To fit into the bigger project it expects the OBJDIR in the environment
(export it in your bigger projects makefile before recursively calling
this makefile).

If you invoke make just as it is, the sourcedir is used as OBJDIR.

The makefile supports 'make V=1' for verbose mode.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>